[Refactor] Fragment package Refactoring - 역할에 따른 package refactoring
