+++ app/src/main/res/layout/activity_home.xml
@@ -17,7 +17,7 @@
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/container_home"
-            android:name="org.sopt.sample.presentation.home.HomeFragment"
+            android:name="org.sopt.sample.presentation.home.fragment.HomeFragment"
             android:layout_width="match_parent"
             android:layout_height="0dp"
             app:layout_constraintBottom_toTopOf="@+id/nav_bottom_home"
