+++ app/src/main/res/layout/fragment_search.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.home.SearchFragment">
+    tools:context=".presentation.home.fragment.SearchFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
